General fixes

+++ Bug Hunting/subdomain_enum.sh
@@ -8,17 +8,19 @@ fi
 if [[ ! -f "resolvers.txt" ]]; then
 	wget https://raw.githubusercontent.com/blechschmidt/massdns/master/lists/resolvers.txt
 fi
-
-#Create the dicctionaries
-
-wget https://raw.githubusercontent.com/danielmiessler/SecLists/master/Discovery/DNS/sortedcombined-knock-dnsrecon-fierce-reconng.txt
-wget https://gist.githubusercontent.com/jhaddix/f64c97d0863a78454e44c2f7119c2a6a/raw/96f4e51d96b2203f19f6381c8c545b278eaa0837/all.txt
-cat sortedcombined-knock-dnsrecon-fierce-reconng.txt >> dictionaries/temporal_subdomains_dicc.txt
-cat all.txt >> dictionaries/temporal_subdomains_dicc.txt
-cat dictionaries/temporal_subdomains_dicc.txt | sort | uniq > dictionaries/subdomains_dicc.txt
-rm sortedcombined-knock-dnsrecon-fierce-reconng.txt
-rm all.txt
-rm dictionaries/temporal_subdomains_dicc.txt
+if [[ ! -d "dictionaries" ]]; then
+	mkdir dictionaries
+fi
+if [[ ! -f "dictionaries/subdomains_dicc.txt" ]]; then
+	wget https://raw.githubusercontent.com/danielmiessler/SecLists/master/Discovery/DNS/sortedcombined-knock-dnsrecon-fierce-reconng.txt
+	wget https://gist.githubusercontent.com/jhaddix/f64c97d0863a78454e44c2f7119c2a6a/raw/96f4e51d96b2203f19f6381c8c545b278eaa0837/all.txt
+	cat sortedcombined-knock-dnsrecon-fierce-reconng.txt >> dictionaries/temporal_subdomains_dicc.txt
+	cat all.txt >> dictionaries/temporal_subdomains_dicc.txt
+	cat dictionaries/temporal_subdomains_dicc.txt | sort | uniq > dictionaries/subdomains_dicc.txt
+	rm sortedcombined-knock-dnsrecon-fierce-reconng.txt
+	rm all.txt
+	rm dictionaries/temporal_subdomains_dicc.txt
+fi
 
 
 DOMAINS=$(cat $1)
@@ -46,42 +48,38 @@ for domain in $DOMAINS; do
         echo "subdomain enum - AltDNS starting..." | slackcat -c $SLACKC -s
         ALT_HOSTS=$domain_no_wc-altDNS_hosts.txt
         altdns -i $RESULT_AMASS -w words.txt -r -s result.out -t 20
-		cat result.out | awk -F: '(NR==0){h1=$1;h2=$2;next} {print $1}' > $ALT_HOSTS
-        #cat result.out | awk '{print $1}' | sed 's/.$//' > $ALT_HOSTS
+	cat result.out | awk -F: '(NR==0){h1=$1;h2=$2;next} {print $1}' > $ALT_HOSTS
         rm result.out
         echo "subdomain enum - AltDNS done" | slackcat -c $SLACKC -s
 
         echo "subdomain enum - merging AltDNS + Amass subdomains" | slackcat -c $SLACKC -s
         cat $RESULT_AMASS >> hosts_merge.txt
         cat $ALT_HOSTS >> hosts_merge.txt
-        cat hosts_merge.txt | sort | uniq > FINALRESULT
-		echo "subdomain enum - merging done" | slackcat -c $SLACKC -s
+        cat hosts_merge.txt | sort | uniq > $FINALRESULT
+	echo "subdomain enum - merging done" | slackcat -c $SLACKC -s
 
-		echo "subdomain enum - blacklisting merge..." | slackcat -c $SLACKC -s
-		comm -23 <(sort FINALRESULT) <(sort subdomains-blacklist.txt) > subdomains_blacklisted.txt
-		echo "subdomain enum - merge blacklisted." | slackcat -c $SLACKC -s
+	echo "subdomain enum - blacklisting merge..." | slackcat -c $SLACKC -s
+	comm -23 <(sort $FINALRESULT) <(sort subdomains-blacklist.txt) > subdomains_blacklisted.txt
+	echo "subdomain enum - merge blacklisted." | slackcat -c $SLACKC -s
 
         echo "subdomain enum - remove temporal files" | slackcat -c $SLACKC -s
         rm hosts_merge.txt
         rm $ALT_HOSTS
-		rm $RESULT_AMASS
-		echo "subdomain enum - remove done" | slackcat -c $SLACKC -s
+	rm $RESULT_AMASS
+	echo "subdomain enum - remove done" | slackcat -c $SLACKC -s
 
         echo "subdomain enum - web application scan starting..." | slackcat -c $SLACKC -s
         cat subdomains_blacklisted.txt | aquatone -ports large -threads 7 -chrome-path $CHROME
         echo "subdomain enum - web application scan done" | slackcat -c $SLACKC -s
        	
-		cat aquatone_urls.txt | grep "https:" > subdomains_p.txt
-
-		for subdomain in $(cat aquatone_urls.txt | grep "http:"); do                                                                                                                         
-			STATUSCODE=$(curl  "$subdomain" -o /dev/null -s -w "%{http_code}\n")
-			if  [[ "$STATUSCODE" == "200" ]];
-			then 
-				echo $subdomain >> subdomains_p.txt
-			fi      
-		done	
-       	
-		
+	cat aquatone_urls.txt | grep "https:" > subdomains_p.txt
+	for subdomain in $(cat aquatone_urls.txt | grep "http:"); do                                                                                                                         
+		STATUSCODE=$(curl  "$subdomain" -o /dev/null -s -w "%{http_code}\n")
+		if  [[ "$STATUSCODE" == "200" ]];
+		then 
+			echo $subdomain >> subdomains_p.txt
+		fi      
+	done	
 
         echo "subdomain enum - uploading subdomains file..." | slackcat -c $SLACKC -s
         if [ ! -f $OFILE ]
